implement loading indications, improve layout
