+++ app/routes/_auth.login.tsx
@@ -1,6 +1,6 @@
 import type { LoaderArgs, ActionArgs } from '@remix-run/node';
 import { redirect } from '@remix-run/node';
-import { Form, useActionData } from '@remix-run/react';
+import { Form, Link, useActionData, useNavigation } from '@remix-run/react';
 import { Card } from '~/components/ui/containers';
 import { Input } from '~/components/ui/forms';
 import { matchesHash } from '~/modules/database/crypto.server';
@@ -47,6 +47,8 @@ export async function loader({ request }: LoaderArgs) {
 
 export default function Component() {
   const actionData = useActionData();
+  const navigation = useNavigation();
+  const isPending = navigation.state === 'submitting' || navigation.state === 'loading';
   return (
     <div className="flex w-full items-center justify-center">
       <div className="max-w-[400px] w-full">
@@ -55,8 +57,13 @@ export default function Component() {
             <h1 className="text-4xl">Log In</h1>
             <Input name="email" type="email" autoComplete="email" placeholder="Email" label="Email" required />
             <Input name="password" type="password" placeholder="Password" label="Password" required />
-            <button type="submit">Log In</button>
+            <button type="submit" disabled={isPending}>
+              {isPending ? 'Logging in...' : 'Log In'}
+            </button>
             {actionData && actionData.message && <p className="text-red-500">{actionData.message}</p>}
+            <p>
+              New here? <Link to="/signup">Sign up</Link>
+            </p>
           </Form>
         </Card>
       </div>
