+++ app/routes/groups.tsx
@@ -1,13 +1,8 @@
-import { Outlet } from '@remix-run/react';
-
 export default function Group() {
   return (
     <div className="flex w-full items-center justify-center mt-20 lg:mt-40">
       <div>
         <h1>List of Groups</h1>
-      </div>
-      <div>
-        <Outlet />
       </div>
     </div>
   );
